Added obstacle sprites

+++ src/player.h
@@ -8,11 +8,11 @@ typedef struct {
     Vector2 acceleration;
 
     Texture2D spritesheet;
-    size_t sprite_width;
-    size_t sprite_height;
+    int sprite_width;
+    int sprite_height;
 
-    size_t hitbox_offset_top;
-    size_t hitbox_offset_bot;
+    int hitbox_offset_top;
+    int hitbox_offset_bot;
 } Player;
 
 void init_player(Player *player, int screen_width, int screen_height);
